backend/requirements.txt
aiofiles
jose
passlib
python_jose
PyYAML
fastapi
uvicorn
gunicorn
python-multipart
aiosqlite
python-magic
httpx
requests
sqlmodel